+++ mrv2/docs/HISTORY.md
@@ -6,6 +6,8 @@ v1.0.8
 - I have enabled TLRENDER_FFMPEG_MINIMAL as a default to compile FFmpeg with
   less muxers, demuxers, parsers and protocols, for smaller file sizes and
   a faster compilation and program start up.
+- Made Help->Documentation point to the online docs if the local docs are not
+  installed.
   
 
 v1.0.7